+++ src/components/Home/Contacts.tsx
@@ -1,11 +1,4 @@
-import {
-  FacebookLogo,
-  GithubLogo,
-  LinkedinLogo,
-  InstagramLogo,
-  YoutubeLogo,
-  TwitterLogo,
-} from "@phosphor-icons/react";
+import { FacebookLogo, GithubLogo, InstagramLogo, LinkedinLogo, TwitterLogo, YoutubeLogo } from "@phosphor-icons/react";
 
 export default function Contacts() {
   const SocialMediaLinks = [
@@ -47,23 +40,19 @@ export default function Contacts() {
         <div className="w-1/2">
           <h2 className="mb-4 text-6xl font-bold">Let's Work Together</h2>
           <p className="mb-11 text-xl">
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
-            eiusmod tempor incididunt ut labore et dolore magna aliqua.
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et
+            dolore magna aliqua.
           </p>
           <button className=" bg-white p-3">
-            <span className="font-body text-[1rem] font-bold uppercase text-[#1E1E1E]">
-              Send a message
-            </span>
+            <span className="font-body text-[1rem] font-bold uppercase text-[#1E1E1E]">Send a message</span>
           </button>
         </div>
         <div className="w-1/2 pl-20">
           <h2 className="mb-4 text-6xl font-bold">Stay Updated</h2>
-          <p className="mb-7 text-xl ">
-            Lorem ipsum dolor sit amet, consectetur.
-          </p>
+          <p className="mb-7 text-xl ">Lorem ipsum dolor sit amet, consectetur.</p>
           <div className="flex space-x-4">
-            {SocialMediaLinks.map(({ link, Icon }) => (
-              <a key={link} href={link} target="_blank" rel="noreferrer">
+            {SocialMediaLinks.map(({ link, Icon }, index) => (
+              <a key={index} href={link} target="_blank" rel="noreferrer">
                 <Icon color="#f8fafc" size={35} weight="fill" />
               </a>
             ))}
