+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
-import Contacts from "@/components/Home/Contacts/Contacts";
-import HeroCarousel from "@/components/Home/Hero/HeroCarousel";
+import Contacts from "@/components/Home/Contacts";
+import HeroCarousel from "@/components/Home/HeroCarousel";
 import LatestNews from "@/components/Home/LatestNews/LatestNews";
-import Mission from "@/components/Home/Mission/Mission";
+import Mission from "@/components/Home/Mission";
 
 /** @TODO: Another carousel below <HeroCarousel />
  *  To track slides: (slide.track.details.rel)
@@ -11,10 +11,7 @@ import Mission from "@/components/Home/Mission/Mission";
 export default function IndexPage() {
   return (
     <>
-      <section className="min-h-screen">
-        <HeroCarousel />
-      </section>
-
+      <HeroCarousel />
       <Mission />
 
       <section className="min-h-screen">
@@ -26,8 +23,8 @@ export default function IndexPage() {
       <section className="min-h-screen">
         <h1 className="text-2xl underline">FAQ</h1>
       </section>
-      {/*CONTACT*/}
-        <Contacts />
+
+      <Contacts />
     </>
   );
 }
